Allow scripts directory to be relative in config file
If the 'scripts' attribute in the 'DEFAULT' section is relative then
convert it to an absolute path relative to wherever the config file is
located.

This allows for config files to be placed in the same directory as the
scripts and just have:

  [DEFAULT]
  scripts = .

Then start the led-manager pointing at this config file and it will
find the scripts.