New library for controlling the Xmas Tree board
New script that uses the new library.

Features of the new library:

  - Primary goal was to see if we could control brightness
    per-led.  This worked but results are not very impressive.
  - Add tree sides and tree layers, though this could easily
    be added to the old library.

There's still a lot of math being done each time the data is
sent to the board, and most of this math will be the same each
time.  We might get better performance if some of the data
was cached.